feat(setup): optional git commands (#8)
* feat(origin): only remove origin if it is template repository

* feat(testSetup): conditionally set git user

+++ scripts/setup.ts
@@ -221,7 +221,10 @@ async function commitAll(message: string) {
 }
 
 async function removeOrigin() {
-  await logAsyncTask(`Removing git origin`, exec(`git remote rm origin`));
+  const { stdout: origin } = await exec(`git remote get-url origin`);
+  if (origin.includes("gjuchault/typescript-library-starter")) {
+    await logAsyncTask(`Removing git origin`, exec(`git remote rm origin`));
+  }
 }
 
 async function logAsyncTask<TResolve>(
